refactor: Make examples executable
Right now, the examples cannot be executed without manually linking NPM
modules to the correct locations, i.e. the `node_modules/` dir in
project root. This makes it really hard to just try out an example as a
user or as a developer working on opencensus-node.

This PR moves the example to a *private* module under `packages/`. This
means that lerna will automatically initialise the necessary symlinks
for developers working on this project. Users on the other hand, could
just run `npm install` and manually execute the examples.