feat: add tracecontext propagation (#97)
Based on w3c TraceContext spec: https://w3c.github.io/distributed-tracing/

The `traceparent` header is validated and propagated according to spec.
The only caveat is that the option flags are lost downstream and reset to
"0x1" by the tracer. This will require a refactor of how the tracer treats
the options as certain assumptions are being made that the least-significant
bit will always represent "sampled".

The `tracestate` is propagated downstream, but is currently not mutable by
the tracer. It is simply stored as a string, but types have been generated
to allow this to be refactored into a more structured type in the future,
allowing the tracer to update the tracestate as necessary.

This commit does not attempt to implement the `Correlation-Context` header.