feat: initial opencensus agent exporter implementation (#128)
* Initial OpenCensus Agent Exporter implementation

The OpenCensus Agent Exporter conforms to the Exporter interface, receiving
spans as they complete, adding them to a buffer, and sending the spans from
the buffer to the agent over a grpc stream.

OpenCensus proto files have been included, as well as types for the required
agent proto files. A series of adapters have been written to adapt from the
@opencensus/core internal span representation to the proto definitions.

Grpc streams are monitored for any disconnect and attempt to reconnect if
possible, utilizing a backoff strategy provided by the grpc client.

The exporter also listens to the Config stream, which recieves configuration
updates from the agent. Specifically, the current tracer sampler is updated
to either a ProbabilitySampler or ConstantSampler. The RateLimitingSampler
option is currently a no-op and has no effect on the tracer sampler.

* Add unit testing to OCAgent Exporter

- Add unit testing of exporter and adapters.

- Ability to pass in a credentials to use for the grpc client connection.

- Move from `grpc-tools` to `google-proto-files`. The `grpc-tools` included
  a lot of unnecessary dependencies, we only need the google proto files.

- Added `remoteParent` to `sameProcessAsParentSpan` property, which flips
  the remoteParent boolean. Note that the `remoteParent` defaults to `false`
  and is NOT being set by any instrumentation today.

- General cleanup.

* Add MockAgent documentation

* Enabled `strictNullChecks` in tsconfig; other nits